Implemented game list cache. The HDD unit has a different implementation from other modes, as scanning the partitions on the HDD already takes way too long. Hence the cache is always used, unless the user manually refreshes the list or renames/deletes a game.

+++ src/hdd.c
@@ -155,7 +155,7 @@ static int hddGetHDLGameInfo(struct GameDataEntry *game, hdl_game_info_t *ginfo)
         ginfo->dma_type = hdl_header->dma_type;
         ginfo->dma_mode = hdl_header->dma_mode;
         ginfo->layer_break = hdl_header->layer1_start;
-        ginfo->disctype = hdl_header->discType;
+        ginfo->disctype = (u8)hdl_header->discType;
         ginfo->start_sector = game->lba;
         ginfo->total_size_in_kb = game->size * 2; //size * 2048 / 1024 = 2x
     } else
